add starlark examples with single and multiple
