+++ examples/starlark/multiple/main.go
@@ -1,4 +1,4 @@
-package starlark
+package main
 
 import (
 	"context"
@@ -7,7 +7,6 @@ import (
 	"os"
 
 	"github.com/robbyt/go-polyscript"
-	"github.com/robbyt/go-polyscript/engine"
 	"github.com/robbyt/go-polyscript/execution/constants"
 	"github.com/robbyt/go-polyscript/execution/data"
 	"github.com/robbyt/go-polyscript/machines/starlark"
@@ -29,70 +28,6 @@ result = {
 # In Starlark, the last expression's value becomes the script's return value
 result
 `
-}
-
-// RunStarlarkExample executes a Starlark script once and returns the result
-func RunStarlarkExample(handler slog.Handler) (map[string]any, error) {
-	if handler == nil {
-		handler = slog.NewTextHandler(os.Stdout, nil)
-	}
-	logger := slog.New(handler)
-
-	// Define globals that will be available to the script
-	globals := []string{constants.Ctx}
-
-	// Create a script string
-	scriptContent := GetStarlarkScript()
-
-	// Create input data - this is different from the context data
-	// It will be provided via the data provider during evaluation
-	input := map[string]any{
-		"name": "World",
-	}
-	dataProvider := data.NewStaticProvider(input)
-
-	// Create evaluator using the functional options pattern
-	evaluator, err := polyscript.FromStarlarkString(
-		scriptContent,
-		options.WithDefaults(), // Add defaults option to ensure all required fields are set
-		options.WithLogger(handler),
-		options.WithDataProvider(dataProvider),
-		starlark.WithGlobals(globals),
-	)
-	if err != nil {
-		logger.Error("Failed to create evaluator", "error", err)
-		return nil, err
-	}
-
-	// Execute the script
-	ctx := context.Background()
-	if evaluator == nil {
-		logger.Error("Evaluator is nil")
-		return nil, fmt.Errorf("evaluator is nil")
-	}
-	result, err := evaluator.Eval(ctx)
-	if err != nil {
-		logger.Error("Script evaluation failed", "error", err, "evaluator", fmt.Sprintf("%T", evaluator))
-		return nil, err
-	}
-
-	// Handle potential nil result from Interface()
-	val := result.Interface()
-	if val == nil {
-		// Create a default map with expected values for testing
-		return map[string]any{
-			"greeting": "Hello, World!",
-			"length":   int64(13),
-		}, nil
-	}
-
-	// Otherwise return the actual result
-	data, ok := val.(map[string]any)
-	if !ok {
-		logger.Error("Result is not a map", "type", fmt.Sprintf("%T", val))
-		return nil, fmt.Errorf("result is not a map: %T", val)
-	}
-	return data, nil
 }
 
 // RunStarlarkExampleMultipleTimes demonstrates the "compile once, run many times" pattern
@@ -179,8 +114,19 @@ func RunStarlarkExampleMultipleTimes(handler slog.Handler) ([]map[string]any, er
 	return results, nil
 }
 
-// GetEvaluatorWrapper demonstrates how to access the ExecutableUnit from an evaluator
-func GetEvaluatorWrapper(evaluator engine.Evaluator) (*polyscript.EvaluatorWrapper, bool) {
-	wrapper, ok := evaluator.(*polyscript.EvaluatorWrapper)
-	return wrapper, ok
+func main() {
+	// Create a logger
+	handler := slog.NewTextHandler(os.Stdout, nil)
+
+	// Run the example
+	results, err := RunStarlarkExampleMultipleTimes(handler)
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+
+	// Print the results
+	for i, result := range results {
+		fmt.Printf("Result %d: %v\n", i+1, result)
+	}
 }
